use np in force function signatures

+++ src/satsim/dynamics/forces.py
@@ -4,13 +4,13 @@ from satsim.environment.earth import Environment
 from satsim.types import StateVector, Vec3
 
 
-def accel_2body(state: StateVector, t: float, env: Environment) -> Vec3:
+def accel_2body(state: StateVector, t: np.float64, env: Environment) -> Vec3:
     # pulls toward Earth center; a = -mu/r^3 * r (r^3 folds in the unit vector)
     r_mag = np.linalg.norm(state.r)
     return -env.mu / r_mag**3 * state.r
 
 
-def accel_j2(state: StateVector, t: float, env: Environment) -> Vec3:
+def accel_j2(state: StateVector, t: np.float64, env: Environment) -> Vec3:
     # Earth bulges at the equator; J2 corrects for the asymmetry
     r = state.r
     r_mag = np.linalg.norm(r)
@@ -26,6 +26,6 @@ def accel_j2(state: StateVector, t: float, env: Environment) -> Vec3:
     )
 
 
-def gravity(state: StateVector, t: float, env: Environment) -> Vec3:
+def gravity(state: StateVector, t: np.float64, env: Environment) -> Vec3:
     # total gravity: two-body + J2 perturbation
     return accel_2body(state, t, env) + accel_j2(state, t, env)
